requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
astroid==2.15.6
async-timeout==4.0.2
attrs==23.1.0
autoflake==2.2.0
backoff==2.2.1
black==23.3.0
blinker==1.6.2
boltons==23.0.0
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.3
colorama==0.4.6
colored==2.2.2
coverage==7.2.7
crashtest==0.4.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.6
dulwich==0.21.5
eliot==1.14.0
eliot-tree==21.0.0
exceptiongroup==1.1.2
faiss-cpu==1.7.4
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
frozenlist==1.3.3
googleapis-common-protos==1.59.1
greenlet==2.0.2
griffe==0.29.1
grpcio==1.58.0
# guard-rails-api-client @ file:///Users/calebcourier/Projects/guardrails-dev/guardrails-poc/guard-rails-api-client
# guardrails-ai @ file:///Users/calebcourier/Projects/guardrails-dev/guardrails
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.0.1
iniconfig==2.0.0
installer==0.7.0
iso8601==2.0.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
keyring==23.13.1
lazy-object-proxy==1.9.0
lockfile==0.12.2
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.1
openai==0.27.8
openapi-python-client==0.14.1
opensearch-py==2.2.0
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-aws-lambda==0.41b0
opentelemetry-instrumentation-dbapi==0.41b0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-grpc==0.41b0
opentelemetry-instrumentation-httpx==0.41b0
opentelemetry-instrumentation-jinja2==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-instrumentation-sqlite3==0.41b0
opentelemetry-instrumentation-tortoiseorm==0.41b0
opentelemetry-instrumentation-urllib==0.41b0
opentelemetry-instrumentation-urllib3==0.41b0
opentelemetry-instrumentation-wsgi==0.41b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
packaging==23.1
pathspec==0.11.1
pep8==1.7.1
pep8-naming==0.13.3
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.8.1
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.23.3
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.4
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-mock==3.11.1
python-dateutil==2.8.2
pytoolconfig==1.2.5
PyYAML==6.0
rapidfuzz==2.15.1
referencing==0.29.1
regex==2023.8.8
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.4.2
rope==1.9.0
rpds-py==0.8.8
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
swagger-ui-py==22.7.13
tenacity==8.2.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
tqdm==4.65.0
trove-classifiers==2023.7.6
typer==0.9.0
typing_extensions==4.6.3
urllib3==1.26.16
virtualenv==20.23.1
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
xattr==0.10.1
yarl==1.9.2
zipp==3.15.0
zope.interface==6.0
